Move InfoChain impls to the engine
