+++ scijava/scijava-ops-api/src/main/java/org/scijava/ops/api/OpInfo.java
@@ -3,12 +3,9 @@ package org.scijava.ops.api;
 
 import java.lang.reflect.AnnotatedElement;
 import java.lang.reflect.Type;
-import java.lang.reflect.TypeVariable;
 import java.util.List;
-import java.util.Map;
 import java.util.stream.Collectors;
 
-import org.scijava.common3.validity.ValidityException;
 import org.scijava.struct.Member;
 import org.scijava.struct.Struct;
 import org.scijava.struct.StructInstance;
@@ -34,7 +31,7 @@ public interface OpInfo extends Comparable<OpInfo> {
 	/** Gets the associated {@link Struct} metadata. */
 	Struct struct();
 
-	/** Gets the hints declared in the {@link OpHints} annotation */
+	/** Gets the hints declared by the Op */
 	Hints declaredHints();
 
 	/** Gets the op's input parameters. */
