+++ scijava/scijava-ops-engine/src/main/java/org/scijava/ops/engine/OpWrapper.java
@@ -1,8 +1,11 @@
-package org.scijava.ops.api;
+package org.scijava.ops.engine;
 
 import java.lang.reflect.ParameterizedType;
 import java.lang.reflect.Type;
 
+import org.scijava.ops.api.OpInstance;
+import org.scijava.ops.api.OpMetadata;
+import org.scijava.ops.api.RichOp;
 import org.scijava.types.Types;
 
 public interface OpWrapper<T> {
