+++ scijava/scijava-ops-engine/src/main/java/org/scijava/ops/engine/impl/OpClassOpInfoGenerator.java
@@ -7,9 +7,9 @@ import java.util.List;
 import org.scijava.common3.validity.ValidityException;
 import org.scijava.meta.Versions;
 import org.scijava.ops.api.Hints;
-import org.scijava.ops.api.OpHints;
+import org.scijava.ops.spi.OpHints;
 import org.scijava.ops.api.OpInfo;
-import org.scijava.ops.api.OpInfoGenerator;
+import org.scijava.ops.engine.OpInfoGenerator;
 import org.scijava.ops.engine.OpUtils;
 import org.scijava.ops.engine.matcher.impl.OpClassInfo;
 import org.scijava.ops.spi.Op;
