+++ scijava/scijava-ops-engine/src/main/java/org/scijava/ops/engine/matcher/adapt/AdaptationInfoChainGenerator.java
@@ -10,7 +10,7 @@ import java.util.List;
 import java.util.Map;
 
 import org.scijava.ops.api.InfoChain;
-import org.scijava.ops.api.InfoChainGenerator;
+import org.scijava.ops.engine.InfoChainGenerator;
 import org.scijava.ops.api.OpInfo;
 import org.scijava.types.Types;
