+++ scijava/scijava-ops-engine/src/main/java/org/scijava/ops/engine/matcher/reduce/ReducedOpInfo.java
@@ -10,7 +10,7 @@ import java.util.stream.Collectors;
 import org.scijava.common3.validity.ValidityException;
 import org.scijava.common3.validity.ValidityProblem;
 import org.scijava.ops.api.Hints;
-import org.scijava.ops.api.OpDescription;
+import org.scijava.ops.engine.OpDescription;
 import org.scijava.ops.api.OpInfo;
 import org.scijava.ops.engine.BaseOpHints;
 import org.scijava.ops.engine.struct.OpResizingMemberParser;
