+++ scijava/scijava-ops-engine/src/main/java/org/scijava/ops/engine/matcher/simplify/Identity.java
@@ -3,7 +3,7 @@ package org.scijava.ops.engine.matcher.simplify;
 
 import java.util.function.Function;
 
-import org.scijava.ops.api.OpHints;
+import org.scijava.ops.spi.OpHints;
 import org.scijava.ops.engine.BaseOpHints.Simplification;
 import org.scijava.ops.spi.Op;
 import org.scijava.ops.spi.OpClass;
