+++ scijava/scijava-ops-engine/src/main/java/org/scijava/ops/engine/matcher/simplify/IdentityLossReporter.java
@@ -1,7 +1,7 @@
 
 package org.scijava.ops.engine.matcher.simplify;
 
-import org.scijava.ops.api.OpHints;
+import org.scijava.ops.spi.OpHints;
 import org.scijava.ops.engine.BaseOpHints.Simplification;
 import org.scijava.ops.engine.conversionLoss.LossReporter;
 import org.scijava.ops.spi.Op;
