+++ scijava/scijava-ops-engine/src/main/java/org/scijava/ops/engine/matcher/simplify/PrimitiveArraySimplifiers.java
@@ -4,7 +4,7 @@ package org.scijava.ops.engine.matcher.simplify;
 import java.util.function.Function;
 
 import org.scijava.collections.ObjectArray;
-import org.scijava.ops.api.OpHints;
+import org.scijava.ops.spi.OpHints;
 import org.scijava.ops.engine.BaseOpHints.Simplification;
 import org.scijava.ops.spi.OpCollection;
 import org.scijava.ops.spi.OpField;
