+++ scijava/scijava-ops-engine/src/main/java/org/scijava/ops/engine/matcher/simplify/SimplifiedOpInfo.java
@@ -12,7 +12,7 @@ import org.scijava.common3.validity.ValidityProblem;
 import org.scijava.meta.Versions;
 import org.scijava.ops.api.Hints;
 import org.scijava.ops.api.InfoChain;
-import org.scijava.ops.api.OpDescription;
+import org.scijava.ops.engine.OpDescription;
 import org.scijava.ops.api.OpEnvironment;
 import org.scijava.ops.api.OpInfo;
 import org.scijava.ops.api.OpRetrievalException;
