+++ scijava/scijava-ops-engine/src/main/resources/META-INF/services/org.scijava.ops.engine.OpWrapper
