+++ scijava/scijava-ops-engine/src/test/java/org/scijava/ops/engine/AbstractTestEnvironment.java
@@ -13,11 +13,8 @@ import java.util.ServiceLoader;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
 import org.scijava.discovery.Discoverer;
-import org.scijava.ops.api.InfoChainGenerator;
 import org.scijava.ops.api.OpEnvironment;
 import org.scijava.ops.api.OpInfo;
-import org.scijava.ops.api.OpInfoGenerator;
-import org.scijava.ops.api.OpWrapper;
 import org.scijava.ops.spi.Op;
 import org.scijava.ops.spi.OpCollection;
