+++ scijava/scijava-ops-engine/src/test/java/org/scijava/ops/engine/impl/OpCachingTest.java
@@ -42,11 +42,11 @@ import org.junit.jupiter.api.Test;
 import org.scijava.discovery.Discoverer;
 import org.scijava.discovery.ManualDiscoverer;
 import org.scijava.function.Producer;
-import org.scijava.ops.api.InfoChainGenerator;
 import org.scijava.ops.api.OpEnvironment;
-import org.scijava.ops.api.OpInfoGenerator;
 import org.scijava.ops.api.OpInstance;
-import org.scijava.ops.api.OpWrapper;
+import org.scijava.ops.engine.InfoChainGenerator;
+import org.scijava.ops.engine.OpInfoGenerator;
+import org.scijava.ops.engine.OpWrapper;
 import org.scijava.ops.engine.MatchingConditions;
 import org.scijava.ops.engine.matcher.MatchingRoutine;
 import org.scijava.ops.spi.Op;
