+++ scijava/scijava-ops-spi/src/main/java/org/scijava/ops/spi/OpHints.java
@@ -1,5 +1,5 @@
 
-package org.scijava.ops.api;
+package org.scijava.ops.spi;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
